Move `buffer_removals` to `server` module
Place it alongside other systems.
We don't need it to be a plugin, users can't toggle it.
Also simplifies unit tests.
